added empty list view component

+++ src/components/RequestsComponents/ProcessedRequestRow/ProcessedRequestRow.js
@@ -1,12 +1,14 @@
-import React from 'react'
-import { CiEdit } from 'react-icons/ci'
+import { useState, useEffect } from 'react'
 import { CiTrash } from 'react-icons/ci'
 import { confirmDialog } from 'primereact/confirmdialog'
+import useAuth from '../../../hooks/useAuth'
+import useRoles from '../../../hooks/useRoles'
+import { Tooltip } from 'react-tooltip'
 import './ProcessedRequestRow.css'
 
 const getFormatedDate = (date) => {
   const dateObj = new Date(date)
-  return `${dateObj.getDate()}.${addZero(
+  return `${addZero(dateObj.getDate())}.${addZero(
     dateObj.getMonth() + 1
   )}.${dateObj.getFullYear()} ${addZero(dateObj.getHours())}:${addZero(
     dateObj.getMinutes()
@@ -18,6 +20,17 @@ const addZero = (number) => {
 }
 
 function RequestRow({ request, deleteRequest }) {
+  const [allowDelete, setAllowDelete] = useState(true)
+  const { auth } = useAuth()
+  const { roles } = useRoles()
+
+  useEffect(() => {
+    if (auth.roles?.includes(roles.Admin)) {
+      setAllowDelete(true)
+    } else {
+      setAllowDelete(false)
+    }
+  }, [roles, auth])
   return (
     <tr className='processed-request-row'>
       <td className='processed-request-row-item processed-request-row-date'>
@@ -32,7 +45,13 @@ function RequestRow({ request, deleteRequest }) {
       <td className='processed-request-row-item processed-request-row-handled-by'>
         <p>{request.handledBy || '-'}</p>
       </td>
-      <td className='processed-request-row-item processed-request-row-accepted'>
+      <td
+        className={
+          request.accepted
+            ? 'processed-request-row-item processed-request-row-accepted processed-request-row-printed-green'
+            : 'processed-request-row-item processed-request-row-accepted processed-request-row-printed-red'
+        }
+      >
         <p>{request.accepted ? 'Da' : 'Nu'}</p>
       </td>
       <td className='processed-request-row-item processed-request-row-rejected-reason'>
@@ -40,10 +59,29 @@ function RequestRow({ request, deleteRequest }) {
       </td>
       {/* only for debugging purposes */}
       <td className='processed-request-row-item processed-request-row-buttons'>
+        <Tooltip id={`tooltip-btn-delete-${request.id}`} />
         <button
+          data-tooltip-id={`tooltip-btn-delete-${request.id}`}
+          data-tooltip-content={
+            allowDelete ? 'Șterge cererea' : 'Nu aveți permisiunea'
+          }
+          data-tooltip-place='left'
           className='processed-request-row-button processed-request-row-delete'
+          disabled={!allowDelete}
           onClick={() => {
-            deleteRequest(request.id)
+            confirmDialog({
+              message: `Sunteți sigur că doriți să ștergeți cererea cu id-ul ${request.id}?`,
+              header: 'Confimare ștergere cerere',
+              icon: 'pi pi-trash',
+              defaultFocus: 'reject',
+              acceptClassName: 'p-button-danger',
+              acceptLabel: 'Da',
+              rejectLabel: 'Nu',
+              accept: () => {
+                deleteRequest(request.id)
+              },
+              reject: () => {},
+            })
           }}
         >
           <CiTrash size={23} />
